feat: Add ShowCommentsCubit and CustomCommentsList
This commit adds the ShowCommentsCubit and CustomCommentsList to the project. The ShowCommentsCubit retrieves comments from Firestore and updates the state accordingly. The CustomCommentsList displays the comments in a ListView. The commit also includes minor changes to other files, such as importing the ShowCommentsCubit in comments.dart and adding Firestore to cSpell.words in settings.json.